优惠券表结构修改

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/coupon/give/CouponGiveListConditionVo.java
@@ -14,7 +14,7 @@ import lombok.Data;
 public class CouponGiveListConditionVo {
 	
 	private String 		couponName;
-	private Integer 	leastConsume;
+	private BigDecimal	leastConsume;
 	private BigDecimal 	denomination;
 	private Integer 	validity;
